Began working on permutations, cycles in math.py
Currently buggy. Need to canonicalize permutations and cycles
in the following manner:

    * Cycles should be rotated so that the minimum value is leftmost
    * Permutation should have disjoint cycles
        - This means that intersecting cycles should be combined
    * Permutations should have cycles sorted by the first index of their tuples

This will allow for easy computation and comparison of permutation elements.
Additionally, this will allow for things like the order of the element to be
computed easily (currently the lcm approach is broken).

These will be used by the PermutationGroup class